WIP: Refactoring remove command
